requirements.txt
munch
pyyaml
tqdm
openai
dotenv
faiss-cpu
rich